+++ src/extension.ts
@@ -7,6 +7,7 @@ if (process.env.SENTRY_DSN) {
   initSentry();
 }
 
+import { handleNewOrUpdatedExtensionInstallation } from "./activation/compareVersions";
 import { ConfluentCloudAuthProvider, getAuthProvider } from "./authn/ccloudProvider";
 import { getCCloudAuthSession } from "./authn/utils";
 import { disableCCloudStatusPolling, enableCCloudStatusPolling } from "./ccloudStatus/polling";
@@ -29,6 +30,7 @@ import { registerKafkaClusterCommands } from "./commands/kafkaClusters";
 import { registerOrganizationCommands } from "./commands/organizations";
 import { registerSchemaRegistryCommands } from "./commands/schemaRegistry";
 import { registerSchemaCommands } from "./commands/schemas";
+import { registerSearchCommands } from "./commands/search";
 import { registerSupportCommands } from "./commands/support";
 import { registerTopicCommands } from "./commands/topics";
 import { AUTH_PROVIDER_ID, AUTH_PROVIDER_LABEL, IconNames } from "./constants";
@@ -236,6 +238,7 @@ async function _activateExtension(
     ...registerFlinkComputePoolCommands(),
     ...registerFlinkStatementCommands(),
     ...registerDocumentCommands(),
+    ...registerSearchCommands(),
   ];
   logger.info("Commands registered");
 
@@ -314,6 +317,8 @@ async function _activateExtension(
 
   // one-time cleanup of old log files from before the rotating log file stream was implemented
   cleanupOldLogFiles();
+
+  await handleNewOrUpdatedExtensionInstallation();
 
   // XXX: used for testing; do not remove
   return context;